Improved performance of PrimeNumber algorithms by introducing a static cache of prime numbers that is grown as new primes are checked.

+++ app/src/main/java/com/arredondo/samples/exercises/StringExercises.java
@@ -203,7 +203,7 @@ public class StringExercises
         {
             for (int i = 0; i < 26; i++)
             {
-                sCharToPrimeMap.put((char) ('a' + i), PrimeNumber.getPrime(i));
+                sCharToPrimeMap.put((char) ('a' + i), (int) PrimeNumber.getPrime(i));
             }
         }
         return sCharToPrimeMap;
